implement loadPkcs12 function
