+++ go.mod
@@ -4,8 +4,11 @@ go 1.22.7
 
 require github.com/stretchr/testify v1.9.0
 
+require golang.org/x/crypto v0.11.0 // indirect
+
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
+	software.sslmate.com/src/go-pkcs12 v0.5.0
 )
